Add files via upload
India has the highest number of road fatalities in the world. 
In 2016, for which global figures are available, India accounted for more than a third of global road accident deaths. The World Health Organization says such deaths are under-reported and estimated that in 2016, the figure for India was likely twice as big as that reported by the government. 

In this project, we have performed a detailed statistical data analysis on Road Accidents in India from 2003 to 2016 using various Python Libraries.